requirements.txt
Celery
Django
django-allauth
django-celery-beat 
django-celery-results # currently requires Django 4.0 or older
django-htmx
Faker
python-decouple
redis
numpy
scikit-surprise
psycopg